backend/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
pydantic-settings
email-validator
supabase
python-jose[cryptography]
passlib[bcrypt]
pytest
httpx
psycopg2-binary
alembic
python-multipart
argon2-cffi
pycountry
